requirements.txt
streamlit
sounddevice
vosk
openai
sentence-transformers
faiss-cpu
torch